Scraper plus scraper route created
The Scraper is something I'm proud of. Its a story of redpemption but more than anything Love. I wrote the scraper in about 3 hours after watching a simple puppeteer tutorial that didnt help. However...., getting it to output through a nodejs/express route is something that took multiple hours of research... but after 3 days of doing it (not consistently as im doing other projects, i finally cracked it! It may not be beautiful or amazingly efficient (which i blame puppeteer more than myself) but it works and its my first function delving into the event loop and using async functions!